requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-ckeditor==6.7.0
django-classy-tags==4.1.0
django-js-asset==2.2.0
django-markdown==0.8.4
django-rosetta==0.10.0
django-taggit==5.0.1
Faker==22.5.1
fire==0.5.0
fonttools==4.47.2
idna==3.6
lxml==5.1.0
Markdown==3.5.2
numpy==1.26.3
opencv-python-headless==4.9.0.80
pdf2docx==0.5.8
pillow==10.2.0
polib==1.2.0
PyMuPDF==1.23.19
PyMuPDFb==1.23.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.9.0
urllib3==2.1.0
